requirements.txt
# SILVIA PoC — Dependencies
# Python 3.10+

# Core
flask==3.1.0
anthropic==0.43.0
python-dotenv==1.0.1

# WhatsApp / Twilio
twilio==9.4.0

# Production server (optional)
gunicorn==23.0.0
